+++ react/index.ts
@@ -2,4 +2,6 @@
 // Requires the `three`, `@react-three/fiber` and `@react-three/drei` peers.
 export * from './usePopulationExpand';
 export * from './ExpandablePopulation';
+export * from './ExpandableNeurons';
+export * from './neuronShaders';
 export * from './VizSpecRenderer';